+++ source/Core/UtilsView.php
@@ -272,21 +272,6 @@ class UtilsView extends \OxidEsales\Eshop\Core\Base
     }
 
     /**
-     * Get template compile id.
-     *
-     * @deprecated since 6.13.0 please define in TemplateRendererBridgeInterface
-     *
-     * @return string
-     */
-    public function getTemplateCompileId()
-    {
-        $shopId = Registry::getConfig()->getShopId();
-        $templateDirectories = $this->getTemplateDirs();
-
-        return md5(reset($templateDirectories) . '__' . $shopId);
-    }
-
-    /**
      * Returns a full path to Smarty compile dir
      *
      * @deprecated since 6.13.0 please use
@@ -330,7 +315,6 @@ class UtilsView extends \OxidEsales\Eshop\Core\Base
         $smarty->compile_dir = $smartyDir;
         $smarty->cache_dir = $smartyDir;
         $smarty->template_dir = $this->getTemplateDirs();
-        $smarty->compile_id = $this->getTemplateCompileId();
 
         $smarty->plugins_dir = array_merge(
             $this->getSmartyPluginDirectories(),