+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `Core\Field`
     - `convertToFormattedDbDate()`
     - `convertToPseudoHtml()`
+  - `OxidEsales\EshopCommunity\Core\UtilsView:getTemplateCompileId()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:parseThroughSmarty()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:_smartyDefaultTemplateHandler()`
   - `oxfunctions:ox_get_template()`